requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==5.0.5
    # via kombu
appy==1.0.7
    # via django-appypod
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via fiona
beautifulsoup4==4.9.3
    # via geotrek
billiard==3.6.3.0
    # via celery
btrees==4.8.0
    # via zodb
cairocffi==1.2.0
    # via
    #   cairosvg
    #   geotrek
    #   weasyprint
cairosvg==2.5.1
    # via weasyprint
celery[redis]==5.0.5
    # via
    #   django-celery-results
    #   geotrek
certifi==2020.12.5
    # via
    #   fiona
    #   requests
    #   sentry-sdk
cffi==1.14.4
    # via
    #   cairocffi
    #   cryptography
    #   persistent
    #   weasyprint
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   fiona
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   fiona
cligj==0.7.1
    # via fiona
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
coverage==5.4
    # via -r requirements.in
cryptography==3.3.1
    # via
    #   pyopenssl
    #   requests
cssselect2==0.4.1
    # via
    #   cairosvg
    #   weasyprint
datetime==4.3
    # via appy
defusedxml==0.6.0
    # via cairosvg
django-appconf==1.0.4
    # via django-compressor
django-appypod==2.0.7
    # via geotrek
django-autocomplete-light[gfk]==3.8.2
    # via -r requirements.in
django-celery-results==2.0.1
    # via geotrek
django-colorfield==0.4.1
    # via geotrek
django-compressor==2.4
    # via
    #   -r requirements.in
    #   geotrek
django-crispy-forms==1.11.0
    # via geotrek
django-debug-toolbar==3.2
    # via -r requirements.in
django-embed-video==1.4.0
    # via
    #   geotrek
    #   paperclip
django-extended-choices==1.3.3
    # via geotrek
django-filter==2.4.0
    # via geotrek
django-geojson==3.1.0
    # via geotrek
django-js-asset==1.2.2
    # via django-mptt
git+https://github.com/GeotrekCE/django-leaflet.git@0.19+geotrek8#egg=django-leaflet
    # via
    #   -r requirements.in
    #   geotrek
django-modelcluster==5.1
    # via geotrek
django-modeltranslation==0.16.1
    # via geotrek
django-mptt==0.11.0
    # via geotrek
django-querysetsequence==0.14
    # via django-autocomplete-light
django-tinymce==3.2.0
    # via geotrek
django-weasyprint==1.0.2
    # via geotrek
django==3.1.6
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-appypod
    #   django-debug-toolbar
    #   django-embed-video
    #   django-filter
    #   django-geojson
    #   django-leaflet
    #   django-modeltranslation
    #   django-mptt
    #   django-querysetsequence
    #   django-weasyprint
    #   djangorestframework
    #   drf-yasg
    #   easy-thumbnails
    #   geotrek
    #   paperclip
djangorestframework-gis==0.17
    # via geotrek
djangorestframework==3.12.2
    # via
    #   djangorestframework-gis
    #   drf-extensions
    #   drf-yasg
    #   geotrek
docutils==0.16
    # via geotrek
drf-dynamic-fields==0.3.1
    # via geotrek
drf-extensions==0.6.0
    # via geotrek
drf-yasg==1.20.0
    # via geotrek
easy-thumbnails==2.7.1
    # via
    #   geotrek
    #   paperclip
factory-boy==3.2.0
    # via
    #   -r requirements.in
    #   geotrek
faker==5.8.0
    # via factory-boy
fiona==1.8.18
    # via geotrek
flake8==3.8.4
    # via -r requirements.in
freezegun==1.1.0
    # via -r requirements.in
geojson==2.5.0
    # via
    #   geotrek
    #   tif2geojson
git+https://github.com/GeotrekCE/Geotrek-admin.git@2.61.1#egg=geotrek
    # via -r requirements.in
gpxpy==1.4.2
    # via geotrek
gunicorn==20.0.4
    # via -r requirements.in
html5lib==1.1
    # via weasyprint
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
kombu==5.0.2
    # via celery
landez==2.5.0
    # via geotrek
lxml==4.6.2
    # via geotrek
markupsafe==1.1.1
    # via jinja2
mbutil==0.3.0
    # via landez
mccabe==0.6.1
    # via flake8
munch==2.5.0
    # via fiona
netifaces==0.10.9
    # via geotrek
packaging==20.9
    # via drf-yasg
paperclip==2.3.2
    # via
    #   -r requirements.in
    #   geotrek
papersize==1.0.1
    # via pdfimpose
pdfimpose==1.1.0
    # via geotrek
persistent==4.7.0
    # via
    #   btrees
    #   zodb
pillow==8.1.0
    # via
    #   cairosvg
    #   easy-thumbnails
    #   geotrek
    #   paperclip
    #   weasyprint
prompt-toolkit==3.0.14
    # via click-repl
psycopg2==2.8.6
    # via
    #   -r requirements.in
    #   geotrek
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pygal==2.4.0
    # via geotrek
pyopenssl==20.0.1
    # via requests
pyparsing==2.4.7
    # via packaging
pypdf2==1.26.0
    # via pdfimpose
pyphen==0.10.0
    # via weasyprint
python-dateutil==2.8.1
    # via
    #   faker
    #   freezegun
python-memcached==1.59
    # via -r requirements.in
pytz==2021.1
    # via
    #   celery
    #   datetime
    #   django
    #   django-modelcluster
rcssmin==1.0.6
    # via django-compressor
redis==3.5.3
    # via celery
requests[security]==2.25.1
    # via
    #   coreapi
    #   django-embed-video
    #   geotrek
    #   landez
rjsmin==1.1.0
    # via django-compressor
ruamel.yaml==0.16.12
    # via drf-yasg
sentry-sdk==0.19.5
    # via -r requirements.in
simplekml==1.3.5
    # via geotrek
six==1.15.0
    # via
    #   click-repl
    #   cryptography
    #   django-autocomplete-light
    #   django-compressor
    #   django-extended-choices
    #   django-modeltranslation
    #   fiona
    #   html5lib
    #   munch
    #   pyopenssl
    #   python-dateutil
    #   python-memcached
    #   zodb
soupsieve==2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
tif2geojson==0.1.3
    # via geotrek
tinycss2==1.1.0
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
transaction==3.0.1
    # via zodb
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.3
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
weasyprint==52.2
    # via
    #   django-weasyprint
    #   geotrek
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
xlrd==2.0.1
    # via geotrek
xmltodict==0.12.0
    # via tif2geojson
zc.lockfile==2.0
    # via zodb
zconfig==3.5.0
    # via zodb
zodb==5.6.0
    # via appy
zodbpickle==2.0.0
    # via zodb
zope.interface==5.4.0
    # via
    #   btrees
    #   datetime
    #   persistent
    #   transaction
    #   zodb

# The following packages are considered to be unsafe in a requirements file:
# setuptools
